Updated Licenses
Apparently I forgot to do it earlier. Basically like other things the libraries
that I'm going to be comitting will be licensed under the AGPLv3/LGPLv3 but
all other code will be AGPLv3 as it is elsewhere to keep the code as copy-left
as possible whilst still letting simple libraries be spread far and wide.

The data(as in the questions) will be CC-BY-ND-NC most likely as I want to
permit non-commerical use but not someone making a profit from my own work
when it comes to the questions themselves and the answers.