Added idb and modified readme.md
I have added the idb library so that I can avoid callback hell since mozilla
thinks it's fine that the program is async and the only way to fully utilize
it is by going through callback hell. No way to have your execution wait on
the results with the default functions.