Added basic db library wrappers & modified ux designs.
Wrote some idb wrapper functions so that in the future if I decide to do an
sql-like syntax for writing queries instead of writing functions to do it.
Also selections will have -1 as the default, and the one to add new things
will be -127. This will be the sole use of negative values within the system.